♻️ refactor: extract column mapping logic into dedicated ColumnMapper module
Move column configuration mapping functions from Frontend to new ColumnMapper module to reduce cognitive load and improve code organization.

- Create packages/cipherstash-proxy/src/postgresql/mapping.rs with ColumnMapper struct
- Move get_projection_columns, get_param_columns, get_literal_columns, and get_column functions
- Update interface to use static methods with dependency injection pattern
- Clean up unused imports in frontend.rs (Identifier, TableColumn, Type)
- Preserve all functionality including logging and error handling
